CapitalizeWord function moved to Helper class in utils package and tested
