--- app/src/main/java/com/example/campusguide/utils/CapitalizeWords.kt
@@ -1,7 +1,0 @@
-package com.example.campusguide.utils
-
-class CapitalizeWords {
-
-
-
-}
