+++ app/src/main/java/com/example/campusguide/utils/Helper.kt
@@ -1,0 +1,16 @@
+package com.example.campusguide.utils
+
+class Helper {
+
+    companion object {
+        fun capitalizeWords(location: String) : String {
+            val words = location.split(" ").toMutableList()
+            var output = ""
+            for (word in words) {
+                output += word.capitalize() + " "
+            }
+            output = output.trim()
+            return output
+        }
+    }
+}
