+++ app/src/test/java/com/example/campusguide/RunAllTestSuite.kt
@@ -3,6 +3,7 @@ package com.example.campusguide
 import com.example.campusguide.directions.CallbackDirectionsConfirmTest
 import com.example.campusguide.directions.DirectionsTest
 import com.example.campusguide.directions.EmptyDirectionsGuardTest
+import com.example.campusguide.directions.RouteTest
 import com.example.campusguide.location.CenterLocationListenerTest
 import com.example.campusguide.location.SwitchCampusTest
 import com.example.campusguide.search.BuildingIndexSingletonTest
@@ -28,7 +29,8 @@ import org.junit.runners.Suite
     SwitchCampusTest::class,
     IndoorSearchResultProviderTest::class,
     IndoorLocationProviderTest::class,
-    BuildingIndexSingletonTest::class
+    BuildingIndexSingletonTest::class,
+    RouteTest::class
 )
 
 class RunAllTestSuite
