+++ app/src/test/java/com/example/campusguide/directions/RouteTest.kt
@@ -1,18 +1,24 @@
 package com.example.campusguide.directions
 
-import com.nhaarman.mockitokotlin2.*
 import org.junit.Test
 
 class RouteTest {
 
-//    @Test
-//    fun testCapitalizeWords() {
-//        var mockWords : List<String> = mock()
-//        val testString = "capitalize the words within this sentence"
-//        whenever(mockWords.
-////        mockWords = testString.split(" ").toMutableList()
-//        assert(mockWords.size == 6)
-//
-//
-//    }
+    @Test
+    fun testCapitalizeWords() {
+
+        val lowerCaseString = "capitalize the words within this sentence"
+        val originalStrings = lowerCaseString.split(" ").toMutableList()
+        var output = ""
+        assert(originalStrings.size == 6)
+
+        for (word in originalStrings) {
+            assert(word.capitalize() != word)
+            val upperCaseWord = word.capitalize()
+            assert(upperCaseWord[0].isUpperCase())
+            output += ("$upperCaseWord ")
+
+        }
+        assert(output.substring(0, output.length - 1) == output.trim())
+    }
 }
